trading_os/requirements.txt
beautifulsoup4
requests
python-dotenv
anthropic
yfinance
